requirements.txt
openai>=1.0.0
rich>=10.0.0
fastapi>=0.104.0
uvicorn>=0.23.0
pydantic>=2.4.0
requests>=2.28.0
python-dotenv>=1.0.0
jwt>=2.10.1